Refactor PotentialDuplicateModal by breaking out UI as MergeModal
