+++ src/features/duplicates/components/PotentialDuplicateModal.tsx
@@ -1,31 +1,15 @@
-import {
-  Alert,
-  AlertTitle,
-  Box,
-  Button,
-  Dialog,
-  DialogActions,
-  DialogTitle,
-  useMediaQuery,
-} from '@mui/material';
-import { FC, useState } from 'react';
-import React, { useEffect } from 'react';
+import { FC } from 'react';
+import React from 'react';
 
-import theme from 'theme';
-import FieldSettings from './FieldSettings';
-import messageIds from '../l10n/messageIds';
 import { PotentialDuplicate } from '../store';
-import PotentialDuplicatesLists from './PotentialDuplicatesLists';
 import useDuplicatesMutations from '../hooks/useDuplicatesMutations';
-import useFieldSettings from '../hooks/useFieldSettings';
-import { useMessages } from 'core/i18n';
 import { useNumericRouteParams } from 'core/hooks';
-import { ZetkinPerson } from 'utils/types/zetkin';
+import MergeModal from './MergeModal';
 
 interface Props {
-  potentialDuplicate: PotentialDuplicate;
   onClose: () => void;
   open: boolean;
+  potentialDuplicate: PotentialDuplicate;
 }
 
 const PotentialDuplicateModal: FC<Props> = ({
@@ -34,92 +18,17 @@ const PotentialDuplicateModal: FC<Props> = ({
   onClose,
 }) => {
   const { orgId } = useNumericRouteParams();
-  const fullScreen = useMediaQuery(theme.breakpoints.down('md'));
-  const messages = useMessages(messageIds);
   const { mergeDuplicate } = useDuplicatesMutations(orgId);
 
-  const [selectedIds, setSelectedIds] = useState<number[]>(
-    potentialDuplicate?.duplicates.map((person) => person.id) ?? []
-  );
-
-  const peopleToMerge = potentialDuplicate?.duplicates.filter((person) =>
-    selectedIds.includes(person.id)
-  );
-
-  const peopleNotToMerge = potentialDuplicate?.duplicates.filter(
-    (person) => !selectedIds.includes(person.id)
-  );
-
-  const { hasConflictingValues, fieldValues, initialOverrides } =
-    useFieldSettings(peopleToMerge);
-  const [overrides, setOverrides] = useState(initialOverrides);
-
-  useEffect(() => {
-    setSelectedIds(
-      potentialDuplicate?.duplicates.map((person) => person.id) ?? []
-    );
-  }, [open]);
-
   return (
-    <Dialog fullScreen={fullScreen} maxWidth={'lg'} open={open}>
-      <DialogTitle sx={{ paddingLeft: 2 }} variant="h5">
-        {messages.modal.title()}
-      </DialogTitle>
-      <Box display="flex" flexGrow={1} overflow="hidden">
-        <Box paddingX={2} sx={{ overflowY: 'auto' }} width="50%">
-          <PotentialDuplicatesLists
-            onDeselect={(person: ZetkinPerson) => {
-              const filteredIds = selectedIds.filter(
-                (item) => item !== person.id
-              );
-              setSelectedIds(filteredIds);
-            }}
-            onSelect={(person: ZetkinPerson) => {
-              const selectedIdsUpdated = [...selectedIds, person.id];
-              setSelectedIds(selectedIdsUpdated);
-            }}
-            peopleNotToMerge={peopleNotToMerge}
-            peopleToMerge={peopleToMerge}
-          />
-        </Box>
-        <Box
-          display="flex"
-          flexDirection="column"
-          marginRight={2}
-          sx={{ overflowY: 'auto' }}
-          width="50%"
-        >
-          <FieldSettings
-            duplicates={peopleToMerge}
-            fieldValues={fieldValues}
-            onChange={(field, value) => {
-              setOverrides({ ...overrides, [`${field}`]: value });
-            }}
-          />
-          <Box marginBottom={2} />
-          {hasConflictingValues && (
-            <Alert severity="warning">
-              <AlertTitle>{messages.modal.warningTitle()}</AlertTitle>
-              {messages.modal.warningMessage()}
-            </Alert>
-          )}
-        </Box>
-      </Box>
-      <DialogActions sx={{ p: 2 }}>
-        <Button onClick={() => onClose()} variant="text">
-          {messages.modal.cancelButton()}
-        </Button>
-        <Button
-          disabled={selectedIds.length > 1 ? false : true}
-          onClick={() =>
-            mergeDuplicate(potentialDuplicate.id, selectedIds, overrides)
-          }
-          variant="contained"
-        >
-          {messages.modal.mergeButton()}
-        </Button>
-      </DialogActions>
-    </Dialog>
+    <MergeModal
+      onClose={onClose}
+      onMerge={(personIds, overrides) => {
+        mergeDuplicate(potentialDuplicate.id, personIds, overrides);
+      }}
+      open={open}
+      persons={potentialDuplicate.duplicates}
+    />
   );
 };
